Rewrote the router to be reusable
